Move IRepository interfaces into the corresponding repository file for higher cohesion
