--- application/account-management/Core/Authentication/Domain/ILoginProcessRepository.cs
@@ -1,5 +1,0 @@
-using PlatformPlatform.SharedKernel.Entities;
-
-namespace PlatformPlatform.AccountManagement.Core.Authentication.Domain;
-
-public interface ILoginRepository : ICrudRepository<Login, LoginId>;
