--- application/account-management/Core/Signups/Domain/ISignupRepository.cs
@@ -1,9 +1,0 @@
-using PlatformPlatform.AccountManagement.Core.Tenants.Domain;
-using PlatformPlatform.SharedKernel.Entities;
-
-namespace PlatformPlatform.AccountManagement.Core.Signups.Domain;
-
-public interface ISignupRepository : ICrudRepository<Signup, SignupId>
-{
-    Signup[] GetByEmailOrTenantId(TenantId tenantId, string email);
-}
